renamed "rendering" to "protocol"
