+++ core/src/main/java/io/github/scroojalix/npcmanager/NPCMain.java
@@ -12,8 +12,8 @@ import io.github.scroojalix.npcmanager.events.EquipmentEvents;
 import io.github.scroojalix.npcmanager.events.NPCEvents;
 import io.github.scroojalix.npcmanager.npc.equipment.EmptySlots;
 import io.github.scroojalix.npcmanager.npc.equipment.EquipmentInventory;
-import io.github.scroojalix.npcmanager.rendering.NPCManager;
-import io.github.scroojalix.npcmanager.rendering.PacketReader;
+import io.github.scroojalix.npcmanager.protocol.NPCManager;
+import io.github.scroojalix.npcmanager.protocol.PacketReader;
 import io.github.scroojalix.npcmanager.storage.Storage;
 import io.github.scroojalix.npcmanager.storage.StorageFactory;
 import io.github.scroojalix.npcmanager.utils.PluginUtils;
